feat(donator steps and goals): i added the translation to the donator steps and Goals
